divided funtions into proper modules, set up promise all
