+++ javascripts/showcaseMovies.js
@@ -4,7 +4,7 @@ let $ = require('jquery');
 
 
 module.exports.movieObjBuilder = (movieData) => {
-	
+	// this is not working and needs to be fixed.
 		for (let i = 0; i < movieData.length; i++) {
 			let movie = {};
 			movie.id = movieData[i].id;
